worked on suggested changes
