+++ recursive-list-main/src/components/breadcrumbs/breadcrumbs.interfaces.ts
@@ -1,0 +1,8 @@
+export interface Breadcrumb {
+  path: string;
+  name: string;
+}
+
+export interface GetBreadcrumbList {
+  (): Breadcrumb[];
+}
