+++ recursive-list-main/src/components/file/file.interfaces.ts
@@ -1,0 +1,4 @@
+export interface FileProps {
+  path: string;
+  name: string;
+}
