+++ recursive-list-main/src/components/folder/folder.component.tsx
@@ -1,12 +1,7 @@
 import { Box, Link, Paper, Typography } from "@material-ui/core";
 import { Link as RouterLink } from "react-router-dom";
+import { FolderProps } from "./folder.interfaces";
 import { useFolderStyles } from "./folder.styles";
-
-import { List } from "components/list/list.component";
-
-export interface FolderProps {
-  folder: List;
-}
 
 const Folder = ({ folder }: FolderProps) => {
   const classes = useFolderStyles();
@@ -15,11 +10,7 @@ const Folder = ({ folder }: FolderProps) => {
     /* Checks if it is a file or folder and renders accordingly */
     <div className={classes.folderContainer}>
       {folder.isFolder ? (
-        <Link
-          style={{ textDecoration: "none" }}
-          component={RouterLink}
-          to={folder.name}
-        >
+        <Link underline="none" component={RouterLink} to={folder.name}>
           <Paper className={classes.paper}>
             <Typography variant="h6">{folder.name}</Typography>
           </Paper>
