+++ recursive-list-main/src/components/folder/folder.interfaces.ts
@@ -1,0 +1,5 @@
+import { List } from "components/list/list.interfaces";
+
+export interface FolderProps {
+  folder: List;
+}
