+++ recursive-list-main/src/components/folder/folder.test.tsx
@@ -1,12 +1,12 @@
 import { render, screen } from "@testing-library/react";
-import { List } from "components/list/list.component";
+import { List } from "components/list/list.interfaces";
 import { MemoryRouter } from "react-router-dom";
 import Folder from "./folder.component";
 
 let folder: List = { isFolder: true, name: "src" };
 
 describe("Folder", () => {
-  it("should open folder if it is a folder", () => {
+  it("Should open folder if it is a folder", () => {
     render(
       <MemoryRouter>
         <Folder folder={folder} />
@@ -19,7 +19,7 @@ describe("Folder", () => {
     );
   });
 
-  it("should render path link when `path` has value", () => {
+  it("Should render path link when `path` has value", () => {
     folder = {
       path: "https://gist.githubusercontent.com/helfi92/96d4444aa0ed46c5f9060a789d316100/raw/ba0d30a9877ea5cc23c7afcd44505dbc2bab1538/typical-live_backing.log",
       name: "home.log",
