+++ recursive-list-main/src/components/list/list.interfaces.ts
@@ -1,0 +1,11 @@
+export interface List {
+  isFolder: boolean;
+  name: string;
+  children?: List[];
+  path?: string;
+}
+
+export interface ListDataProps {
+  list: List[];
+  parent?: string;
+}
