+++ recursive-list-main/src/utils/params/params.util.ts
@@ -1,9 +1,8 @@
 export const getParamFromPathname = (pathname: string): string => {
   /* returns the current pathname*/
 
-  let param = "";
   const pathnameArray = pathname.split("/");
   const lastIndex = pathnameArray.length - 1;
-  param = pathnameArray[lastIndex].replace("/", "");
+  const param = pathnameArray[lastIndex].replace("/", "");
   return param;
 };
